requirements.txt
django==1.3.1
PIL==1.1.7
South==0.8.3
django-threadedcomments==0.5.3
django-photologue==2.3
django-tagging==0.3.1
django-avatar==1.0.5
django-pagination==1.0.7
django-uni-form==0.9.0
